% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_scxml.R
\name{export_scxml}
\alias{export_scxml}
\title{export_scxml}
\usage{
export_scxml(data, label_names, file_name = "SportsCode.xml", colours = NULL)
}
\arguments{
\item{data}{a data frame or tibble that containes columns ID, start, end and code}

\item{label_names}{a vector of column names that you want to export as labels}

\item{file_name}{a character string to name the exported file}

\item{colours}{optional dataframe or tibble of colour codes to colour instances}
}
\value{
saves an xml file to the current working directory
}
\description{
export_scxml
}
