% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportscode_colours.R
\name{sportscode_colours}
\alias{sportscode_colours}
\title{sportscode_colours}
\usage{
sportscode_colours(filename)
}
\arguments{
\item{filename}{the file path of the xml file to import the colours from}
}
\value{
a tibble of sportscode colours for each instance, to be used in conjunction with export_scxml
}
\description{
sportscode_colours
}
