% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportscode_df.R
\name{sportscode_df}
\alias{sportscode_df}
\title{sportscode_df}
\usage{
sportscode_df(filename, remove_dup = FALSE)
}
\arguments{
\item{filename}{the file path of the xml file to import}

\item{remove_dup}{whether to remove duplicate labels and only retain the first label of each group for each instance. Default is FALSE.}
}
\value{
a tibble of the xml data in wide format, with ID, start, end and code columns intact in a structure required to then export back into a compatible xml file.
}
\description{
sportscode_df
}
